[ALPS06148957] Cast from enum to unsigned int to avoid negative value
Remove the risk of negative value indexing.

MTK-Commit-Id: 3a9eacbaf1b77d8276640227768856913e1f2825

CR-Id: ALPS06148957
Feature: Wi-Fi Link Layer Statistics

+++ os/linux/gl_vendor.c
@@ -923,7 +923,7 @@ void dumpLinkStatsAc(struct STATS_LLS_WMM_AC_STAT *ac_stat,
 		"VO", "VI", "BE", "BK"};
 
 	DBGLOG(REQ, INFO, "AC[%s] %u %u %u %u %u %u %u %u %u %u %u %u %u %u %u",
-			s[ac],
+			s[(uint32_t)ac],
 			ac_stat[ac].ac,
 			ac_stat[ac].tx_mpdu,
 			ac_stat[ac].rx_mpdu,
